requirements.txt
flask
flask-cors
groq
pymupdf
python-docx
python-dotenv